Déplacement de la page Compte et refactorisation avec correction des imports
